+++ openmarket-server/src/main/java/co/unicauca/openmarket/server/access/IUserRepository.java
@@ -8,19 +8,21 @@ package co.unicauca.openmarket.server.access;
 import co.unicauca.openmarket.commons.domain.User;
 
 public interface IUserRepository {
-    
+
     /**
      * Busca un usuario por email y contraseña
-     * @param email email del usuario
+     * 
+     * @param email    email del usuario
      * @param password contraseña del usuario
      * @return retorna el usuario
      */
     public User findByEmailAndPassword(String email, String password);
-    
+
     /**
      * Busca el Email y la Contraseña de un usuario
+     * 
      * @param userId Id del usuario
      * @return retorna el usuario
      */
-    public User findById(int userId);
+    public User findById(Long userId);
 }
